refactor: Error 클래스 export 형태 수정
